# internal/handlers/swagger.json
{"openapi":"3.0.3","info":{"title":"Epoch Days API","version":"1.0.0","description":"API REST para calcular o epoch atual e o epoch futuro com base em uma quantidade de dias.\n"},"servers":[{"url":"http://localhost:8080","description":"Local"},{"url":"https://{service}.onrender.com","description":"Render","variables":{"service":{"default":"epoch-days"}}},{"url":"https://{service}.up.railway.app","description":"Railway","variables":{"service":{"default":"epoch-days"}}}],"paths":{"/epoch/{days}":{"get":{"tags":["epoch"],"operationId":"getEpochByDays","summary":"Retorna epoch atual e epoch futuro somando dias","description":"Recebe um inteiro `days` via path e calcula:\n- `now_epoch`: epoch UTC atual (segundos desde 1970)\n- `future_epoch`: `now_epoch + days * 86400`\n","parameters":[{"name":"days","in":"path","required":true,"description":"Quantidade de dias para somar ao epoch atual (inteiro, pode ser negativo)","schema":{"type":"integer","format":"int32"},"example":5}],"responses":{"200":{"description":"Cálculo realizado com sucesso","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EpochResponse"},"examples":{"success":{"value":{"now_epoch":1700000000,"future_epoch":1700432000,"days_added":5}}}}}},"400":{"description":"Parâmetro days inválido","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"examples":{"invalidDays":{"value":{"error":"invalid days: must be an integer"}}}}}},"408":{"description":"Requisição cancelada/expirada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"examples":{"canceled":{"value":{"error":"request context canceled"}}}}}},"500":{"description":"Erro interno (ex.: overflow)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"examples":{"internal":{"value":{"error":"integer overflow when adding days"}}}}}}}}}},"components":{"schemas":{"EpochResponse":{"type":"object","required":["now_epoch","future_epoch","days_added"],"properties":{"now_epoch":{"type":"integer","format":"int64","description":"Epoch UTC atual em segundos","example":1700000000},"future_epoch":{"type":"integer","format":"int64","description":"Epoch UTC futuro em segundos","example":1700432000},"days_added":{"type":"integer","format":"int32","description":"Valor de dias recebido na URL","example":5}}},"ErrorResponse":{"type":"object","required":["error"],"properties":{"error":{"type":"string","description":"Mensagem de erro","example":"invalid days: must be an integer"}}}}}}